requirements.txt
pip-upgrader
streamlit==1.2.0
torch==1.10.0
torchvision==0.11.1
Pillow==8.4.0
facenet-pytorch == 2.5.2
matplotlib == 3.5.0
numpy == 1.20.3
opencv-python-headless